rename CustomizeSizeActivity to Settings Activity. Created classes for better readability
