+++ app/src/main/res/layout/activity_settings.xml
@@ -1,0 +1,59 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ScrollView
+      android:id="@+id/settings"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" >
+        <LinearLayout
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+            android:orientation="vertical">
+        <RelativeLayout
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:orientation="horizontal"
+          android:id="@+id/subcriber_id_container"
+          android:padding="4dp"
+          android:tint="#828282">
+            <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="订阅ID"
+              android:layout_alignParentStart="true"/>
+
+            <Spinner
+              android:id="@+id/subcriber_id_setter"
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:spinnerMode="dialog"
+              android:layout_alignParentEnd="true"/>
+        </RelativeLayout>
+
+        <RelativeLayout
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:orientation="horizontal"
+          android:id="@+id/time_format_container"
+          android:padding="4dp"
+          android:tint="#828282">
+            <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="时间显示格式"
+              android:layout_alignParentStart="true"/>
+            <Spinner
+              android:id="@+id/time_format_setter"
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:spinnerMode="dropdown"
+              android:layout_alignParentEnd="true"/>
+
+        </RelativeLayout>
+
+        </LinearLayout>
+
+
+    </ScrollView>
+</LinearLayout>
